% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dependency_graph.R
\name{plot_dependency_graph}
\alias{plot_dependency_graph}
\title{Brief: Plot dependencies map from dep_map and save HTML file}
\usage{
plot_dependency_graph(
  dep_map,
  output_path,
  output_name,
  main_node = script_name
)
}
\arguments{
\item{dep_map}{A named list of data frames, where each data frame contains the dependencies of a user-defined function.}

\item{output_path}{path to save function map to (defined in funcMapper)}

\item{output_name}{name of the function map (defined in funcMapper)}

\item{main_node}{this is always set to the script name, generated from script path in funcMapper. Used to highlight main script node in red.}
}
\value{
A visNetwork plot of the user created function map, saved to the output path
}
\description{
Description: This function plots the dep_map from build_dependency_map(), by first passing it through get_edges_from_map()
to convert it from a list of data frames to a unified edge list, which is then used in a visNetwork plot. This is then
saved to the output path with the output name (both defined in funcMapper) as an HTML file.
}
\details{
Author: Antonio Fratamico
Date: 10/07/2025
}
